\name{HVGR-package}
\alias{HVGR-package}
\alias{HVGR}
\docType{package}
\title{
\packageTitle{HVGR}
}
\description{
\packageDescription{HVGR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HVGR}
\packageIndices{HVGR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HVGR}

Maintainer: \packageMaintainer{HVGR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
